+++ hvega/tests/Gallery/Scatter.hs
@@ -150,7 +150,7 @@ scatter7 =
 
         config =
             configure
-                . configuration (RemoveInvalid False)
+                . configuration (MarkStyle [ MRemoveInvalid False ])
 
         enc =
             encoding